pyproject.toml
[tool.poetry]
name = "einsatzleiter"
version = "0.1.0"
description = "Anwendung zur übersichtlichen Abfassung eines Funktagebuches im (Feuerwehr-)Einsatz"
authors = ["Swen Lenk (sw-lnk)"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
customtkinter = "^5.2.2"
darkdetect = "^0.8.0"
defusedxml = "^0.7.1"
dnspython = "^2.6.1"
fonttools = "^4.51.0"
packaging = "^24.0"
pillow = "^10.3.0"
ttkbootstrap = "^1.10.1"
sqlalchemy = "^2.0.29"
sqlalchemy-utils = "^0.41.2"
fpdf2 = "^2.7.8"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
